+++ tests/unit/core/binarycodec/test_main.py
@@ -331,14 +331,14 @@ class TestXAddress(TestCase):
 
 
 class TestMainFixtures(TestCase):
-    maxDiff = 1000
+    maxDiff = None
 
     def _check_binary_and_json(self, test):
         test_binary = test["binary"]
         test_json = test["json"]
         with self.subTest(test_binary=test_binary, test_json=test_json):
-            self.assertEqual(encode(test_json), test_binary)
             self.assertEqual(decode(test_binary), test_json)
+            self.assertEqual(encode(test_json), test_binary)
 
     def _check_xaddress_jsons(self, test):
         x_json = test["xjson"]