+++ xrpl/models/transactions/__init__.py
@@ -58,6 +58,15 @@ from xrpl.models.transactions.trust_set import (
     TrustSetFlag,
     TrustSetFlagInterface,
 )
+from xrpl.models.transactions.xchain_account_create_commit import (
+    XChainAccountCreateCommit,
+)
+from xrpl.models.transactions.xchain_add_attestation import XChainAddAttestation
+from xrpl.models.transactions.xchain_claim import XChainClaim
+from xrpl.models.transactions.xchain_commit import XChainCommit
+from xrpl.models.transactions.xchain_create_bridge import XChainCreateBridge
+from xrpl.models.transactions.xchain_create_claim_id import XChainCreateClaimID
+from xrpl.models.transactions.xchain_modify_bridge import XChainModifyBridge
 
 __all__ = [
     "AccountDelete",
@@ -109,4 +118,11 @@ __all__ = [
     "TrustSet",
     "TrustSetFlag",
     "TrustSetFlagInterface",
+    "XChainAddAttestation",
+    "XChainClaim",
+    "XChainCommit",
+    "XChainCreateBridge",
+    "XChainCreateClaimID",
+    "XChainModifyBridge",
+    "XChainAccountCreateCommit",
 ]